pyproject.toml
[project]
name = "aisandbox"
version = "0.1.0"
description = "testy mc AI face"
authors = [
    {name = "BRL"}
]
license = {text = "free to use"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai (>=0.8.3,<0.9.0)",
    "google-api-python-client (>=2.157.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "openai (>=1.59.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
